\name{HiTC-internal}
\alias{myPalette}
\alias{getDeltaRange}
\alias{tricube}
\alias{gdiag}
\alias{addImageTracks}
\alias{heatmapC}
\alias{triViewC}
\alias{pair.chrom}

\title{Internal HiTC functions}
\description{
  Called internally by other \code{HiTC} functions.
  Normally need not be called by the user.
}
\author{Nicolas Servant}
\seealso{\code{\link{HTCexp-class}}}
\keyword{internal}
